Moved the download template views out of the API realm into the main realm.
