+++ biosys/apps/main/api/urls.py
@@ -3,44 +3,36 @@ from __future__ import absolute_import, unicode_literals, print_function, divisi
 from django.conf.urls import url
 from rest_framework import routers
 
-from main.api import views as main_views
+from main.api import views as api_views
 
 router = routers.DefaultRouter()
-router.register(r'users?', main_views.UserViewSet, 'user')
-router.register(r'projects?', main_views.ProjectViewSet, 'project')
-router.register(r'sites?', main_views.SiteViewSet, 'site')
-router.register(r'datasets?', main_views.DatasetViewSet, 'dataset')
-router.register(r'records?', main_views.RecordViewSet, 'record')
+router.register(r'users?', api_views.UserViewSet, 'user')
+router.register(r'projects?', api_views.ProjectViewSet, 'project')
+router.register(r'sites?', api_views.SiteViewSet, 'site')
+router.register(r'datasets?', api_views.DatasetViewSet, 'dataset')
+router.register(r'records?', api_views.RecordViewSet, 'record')
 
 url_patterns = [
-    url(r'projects?/(?P<pk>\d+)/sites/?', main_views.ProjectSitesView.as_view(), name='project-sites'),  # bulk sites
-    url(r'projects?/(?P<pk>\d+)/upload-sites/?', main_views.ProjectSitesUploadView.as_view(),
+    url(r'projects?/(?P<pk>\d+)/sites/?', api_views.ProjectSitesView.as_view(), name='project-sites'),  # bulk sites
+    url(r'projects?/(?P<pk>\d+)/upload-sites/?', api_views.ProjectSitesUploadView.as_view(),
         name='upload-sites'),  # file upload for sites
-    url(r'datasets?/(?P<pk>\d+)/records/?', main_views.DatasetRecordsView.as_view(), name='dataset-records'),
+    url(r'datasets?/(?P<pk>\d+)/records/?', api_views.DatasetRecordsView.as_view(), name='dataset-records'),
     # upload data files
-    url(r'datasets?/(?P<pk>\d+)/upload-records/?', main_views.DatasetUploadRecordsView.as_view(),
+    url(r'datasets?/(?P<pk>\d+)/upload-records/?', api_views.DatasetUploadRecordsView.as_view(),
         name='dataset-upload'),
-    url(r'statistics/?', main_views.StatisticsView.as_view(), name="statistics"),
-    url(r'whoami/?', main_views.WhoamiView.as_view(), name="whoami"),
-    url(r'species/?', main_views.SpeciesView.as_view(), name="species"),
-    url(r'logout/?', main_views.LogoutView.as_view(), name="logout"),
+    url(r'statistics/?', api_views.StatisticsView.as_view(), name="statistics"),
+    url(r'whoami/?', api_views.WhoamiView.as_view(), name="whoami"),
+    url(r'species/?', api_views.SpeciesView.as_view(), name="species"),
+    url(r'logout/?', api_views.LogoutView.as_view(), name="logout"),
     # utils
     url(r'utils/geometry-to-data/record/(?P<pk>\d+)/?',
-        main_views.GeoConvertView.as_view(output='data'),
+        api_views.GeoConvertView.as_view(output='data'),
         name="geometry-to-data"
         ),
     url(r'utils/data-to-geometry/record/(?P<pk>\d+)/?',
-        main_views.GeoConvertView.as_view(output='geometry'),
+        api_views.GeoConvertView.as_view(output='geometry'),
         name="data-to-geometry"
-        ),
-    url(r'utils/templates/site/lat-long/?',
-        main_views.SiteTemplateView.as_view(model='lat_long'),
-        name="site-template-lat-long"
-        ),
-    url(r'utils/templates/site/easting-northing/?',
-        main_views.SiteTemplateView.as_view(model='easting_northing'),
-        name="site-template-easting-northing"
-        ),
+        )
 ]
 
 urls = router.urls + url_patterns
