+++ biosys/apps/main/tests/api/test_site.py
@@ -505,65 +505,3 @@ class TestSerialization(helpers.BaseUserTestCase):
         centroid = GEOSGeometry(json.dumps(data['centroid']))
         self.assertEqual(centroid, site.geometry.centroid)
 
-
-class TestDownloadTemplates(helpers.BaseUserTestCase):
-
-    def test_lat_long(self):
-        client = self.custodian_1_client
-        url = reverse('api:site-template-lat-long')
-        resp = client.get(url)
-        self.assertEqual(resp.status_code, status.HTTP_200_OK)
-        self.assertEqual(resp.get('content-type'),
-                         'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
-        content_disposition = resp.get('content-disposition')
-        # should be something like:
-        # 'attachment; filename=Sites_template_lat_long.xlsx
-        match = re.match('attachment; filename=(.+)', content_disposition)
-        self.assertIsNotNone(match)
-        filename, ext = path.splitext(match.group(1))
-        self.assertEquals(ext, '.xlsx')
-        self.assertEquals(filename, 'Sites_template_lat_long')
-        # read content
-        wb = load_workbook(six.BytesIO(resp.content), read_only=True)
-        # one datasheet named 'Sites'
-        expected_sheet_name = 'Sites'
-        sheet_names = wb.get_sheet_names()
-        self.assertEquals(1, len(sheet_names))
-        self.assertEquals(sheet_names[0], expected_sheet_name)
-        ws = wb.get_sheet_by_name(expected_sheet_name)
-        rows = list(ws.rows)
-        # only one row
-        self.assertEquals(len(rows), 1)
-        got_headers = [c.value for c in rows[0]]
-        expected_headers = ['Name', 'Code', 'Description', 'Latitude', 'Longitude', 'Datum']
-        self.assertEqual(got_headers, expected_headers)
-
-    def test_easting_northing(self):
-        client = self.custodian_1_client
-        url = reverse('api:site-template-easting-northing')
-        resp = client.get(url)
-        self.assertEqual(resp.status_code, status.HTTP_200_OK)
-        self.assertEqual(resp.get('content-type'),
-                         'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
-        content_disposition = resp.get('content-disposition')
-        # should be something like:
-        # 'attachment; filename=Sites_template_lat_long.xlsx
-        match = re.match('attachment; filename=(.+)', content_disposition)
-        self.assertIsNotNone(match)
-        filename, ext = path.splitext(match.group(1))
-        self.assertEquals(ext, '.xlsx')
-        self.assertEquals(filename, 'Sites_template_easting_northing')
-        # read content
-        wb = load_workbook(six.BytesIO(resp.content), read_only=True)
-        # one datasheet named 'Sites'
-        expected_sheet_name = 'Sites'
-        sheet_names = wb.get_sheet_names()
-        self.assertEquals(1, len(sheet_names))
-        self.assertEquals(sheet_names[0], expected_sheet_name)
-        ws = wb.get_sheet_by_name(expected_sheet_name)
-        rows = list(ws.rows)
-        # only one row
-        self.assertEquals(len(rows), 1)
-        got_headers = [c.value for c in rows[0]]
-        expected_headers = ['Name', 'Code', 'Description', 'Easting', 'Northing', 'Datum', 'Zone']
-        self.assertEqual(got_headers, expected_headers)
